# Support HTTP/2 with TLS in OTLP HTTP receiver.