--- a/config/confighttp/confighttp_test.go
+++ b/config/confighttp/confighttp_test.go
@@ -380,6 +380,7 @@ func TestHttpReception(t *testing.T) {
 		tlsServerCreds *configtls.TLSServerSetting
 		tlsClientCreds *configtls.TLSClientSetting
 		hasError       bool
+		forceHttp1     bool
 	}{
 		{
 			name:           "noTLS",
@@ -403,6 +404,23 @@ func TestHttpReception(t *testing.T) {
 				},
 				ServerName: "localhost",
 			},
+		},
+		{
+			name: "TLS (HTTP/1.1)",
+			tlsServerCreds: &configtls.TLSServerSetting{
+				TLSSetting: configtls.TLSSetting{
+					CAFile:   filepath.Join("testdata", "ca.crt"),
+					CertFile: filepath.Join("testdata", "server.crt"),
+					KeyFile:  filepath.Join("testdata", "server.key"),
+				},
+			},
+			tlsClientCreds: &configtls.TLSClientSetting{
+				TLSSetting: configtls.TLSSetting{
+					CAFile: filepath.Join("testdata", "ca.crt"),
+				},
+				ServerName: "localhost",
+			},
+			forceHttp1: true,
 		},
 		{
 			name: "NoServerCertificates",
@@ -505,15 +523,24 @@ func TestHttpReception(t *testing.T) {
 			<-time.After(10 * time.Millisecond)
 
 			prefix := "https://"
+			expectedProto := "HTTP/2.0"
 			if tt.tlsClientCreds.Insecure {
 				prefix = "http://"
+				expectedProto = "HTTP/1.1"
 			}
 
 			hcs := &HTTPClientSettings{
 				Endpoint:   prefix + ln.Addr().String(),
 				TLSSetting: *tt.tlsClientCreds,
 			}
-			client, errClient := hcs.ToClient(map[config.ComponentID]component.Extension{}, componenttest.NewNopTelemetrySettings())
+			if tt.forceHttp1 {
+				expectedProto = "HTTP/1.1"
+				hcs.CustomRoundTripper = func(rt http.RoundTripper) (http.RoundTripper, error) {
+					rt.(*http.Transport).ForceAttemptHTTP2 = false
+					return rt, nil
+				}
+			}
+			client, errClient := hcs.ToClient(map[config.ComponentID]component.Extension{}, component.TelemetrySettings{})
 			require.NoError(t, errClient)
 
 			resp, errResp := client.Get(hcs.Endpoint)
@@ -524,6 +551,7 @@ func TestHttpReception(t *testing.T) {
 				body, errRead := ioutil.ReadAll(resp.Body)
 				assert.NoError(t, errRead)
 				assert.Equal(t, "test", string(body))
+				assert.Equal(t, expectedProto, resp.Proto)
 			}
 			require.NoError(t, s.Close())
 		})