Currently the HTTP receiver never negotiates HTTP/2 even when the client supports it. In a typical scenario of having many apps with batch exporters hitting a collector, HTTP/1 and HTTP/2 should perform mostly the same, but if putting a load balancer in front of the collector, HTTP/2 will be able to multiplex many apps across a single connection and can perform quite a bit better.

Benchmarks are noisy since executing actual HTTP requests. Hopefully we can see some trends

HTTP/2 seems to provide about 3x performance when using a shared client (behaves like a load balancer)
HTTP/2 performs slightly slower with a client per thread (behaves like an app), but this is probably within noise. It is something to confirm we're OK with - I think most services now adays will support HTTP/2 though even if it can be slightly slower in non-shared cases. One way to think of it is gRPC is HTTP/2-only anyways and we'd expect HTTP receiver to have a similar baseline.

After
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/collector/config/confighttp
cpu: Intel(R) Xeon(R) Platinum 8175M CPU @ 2.50GHz
BenchmarkHttpRequest
BenchmarkHttpRequest/HTTP/2.0,_shared_client_(like_load_balancer)
BenchmarkHttpRequest/HTTP/2.0,_shared_client_(like_load_balancer)-16         	   29907	     39563 ns/op
BenchmarkHttpRequest/HTTP/1.1,_shared_client_(like_load_balancer)
BenchmarkHttpRequest/HTTP/1.1,_shared_client_(like_load_balancer)-16         	    9814	    110057 ns/op
BenchmarkHttpRequest/HTTP/2.0,_client_per_thread_(like_single_app)
BenchmarkHttpRequest/HTTP/2.0,_client_per_thread_(like_single_app)-16        	   27634	     38459 ns/op
BenchmarkHttpRequest/HTTP/1.1,_client_per_thread_(like_single_app)
BenchmarkHttpRequest/HTTP/1.1,_client_per_thread_(like_single_app)-16        	   15918	     72502 ns/op
PASS
```

Before (HTTP/2 is actually 1.1)
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/collector/config/confighttp
cpu: Intel(R) Xeon(R) Platinum 8175M CPU @ 2.50GHz
BenchmarkHttpRequest
BenchmarkHttpRequest/HTTP/2.0,_shared_client_(like_load_balancer)
BenchmarkHttpRequest/HTTP/2.0,_shared_client_(like_load_balancer)-16         	    9368	    116053 ns/op
BenchmarkHttpRequest/HTTP/1.1,_shared_client_(like_load_balancer)
BenchmarkHttpRequest/HTTP/1.1,_shared_client_(like_load_balancer)-16         	   10587	    113067 ns/op
BenchmarkHttpRequest/HTTP/2.0,_client_per_thread_(like_single_app)
BenchmarkHttpRequest/HTTP/2.0,_client_per_thread_(like_single_app)-16        	   13497	     81537 ns/op
BenchmarkHttpRequest/HTTP/1.1,_client_per_thread_(like_single_app)
BenchmarkHttpRequest/HTTP/1.1,_client_per_thread_(like_single_app)-16        	   14283	     76041 ns/op
PASS

```